Try splitting things differently
